beta

+++ src/components/account/blog.vue
@@ -83,17 +83,19 @@
         <div class="input-group mt-3">
             <textarea name="" id="" cols="5" rows="5" class="form-control" v-model="Form.desc" placeholder="description" required></textarea>
         </div>
-        <div class="input-group mt-3">
-          <div class="border border-1 border-secondary p-5 rounded-3 " v-if="previewImage === ''" @click="imageUploadUi">
-            <i class="fas fa-plus"></i>
+        <div class="input-group mt-3 d-flex justify-content-center align-items-center ">
+          <div class="upload-img-ui" v-if="previewImage === ''" @click="imageUploadUi">
+            <img src="../../assets/img/icons/image-default.png"  alt="">
           </div>
           <input type="file" id="imageUpload" @change="imageUpload" class="form-control d-none">
         </div>
-        <div class="input-group mt-3" v-if="previewImage !== ''">
-          <img :src="previewImage" class=" img-fluid border border-1 border-light rounded-2 shadow-sm" @click="imageUploadUi" height="50px" alt="">
+        <div class="input-group mt-3 d-flex justify-content-center align-items-center" v-if="previewImage !== ''">
+          <img :src="previewImage" class=" img-fluid border border-1 border-light rounded-2 shadow-sm " @click="imageUploadUi" alt="">
         </div>
         <div class="input-group mt-3">
-          <button class="btn btn-outline-dark ">{{ is_edit === true ? 'Update' : 'Create' }}</button>
+          <button class="btn btn-outline-dark form-control btn-sm" :disabled="spinner">
+            <span v-if="spinner" class="spinner-border spinner-border-sm" role="status" aria-hidden="true"></span>
+            {{ is_edit === true ? 'Update' : 'Create' }}</button>
         </div>
       </form>
     </b-modal>
@@ -111,6 +113,7 @@ import uniqid from 'uniqid';
 export default {
         data() {
             return {
+
                 Form:{
                     'title' : '',
                     'image' : '',
@@ -121,7 +124,8 @@ export default {
                 file : '',
                 previewImage : '',
                 is_edit : false,
-                data : []
+                data : [],
+                spinner : false,
             }
         },
         created() {
@@ -152,6 +156,7 @@ export default {
               fileReader.readAsDataURL(files[0]);
             },
             formatImage() {
+              this.spinner = true;
               // Upload To Firebase Storage
               const storage = getStorage();
               const storageRef = ref(storage, "blog/" + uniqid() + this.fileName);
@@ -224,15 +229,23 @@ export default {
               }).then(el => {
                 this.$bvModal.hide("modal-blog");
                 this.getData();
+                this.spinner = false;
               });
             },
             async del(d) {
-              const ColRef = collection(db, "blog");
-              let Ref = doc(ColRef, d.id);
-              await deleteDoc(Ref);
-              this.data = this.data.filter((el) => {
-                return el.id !== d.id;
-              });
+              let text = "Are you sure , you want to delete.";
+              if (window.confirm(text) == true) {
+
+                const ColRef = collection(db, "blog");
+                let Ref = doc(ColRef, d.id);
+                await deleteDoc(Ref);
+                this.data = this.data.filter((el) => {
+                  return el.id !== d.id;
+                });
+
+              } else {
+                return;
+              }
             },
             // Crud End
             // Upload To Firebase Store
@@ -243,7 +256,9 @@ export default {
                   'image' : this.Form.image,
                   'desc'  : this.Form.desc,
                   'date'  : this.Form.date,
-                });
+                }).then((el)=>{
+                  this.spinner = false;
+                })
                 this.Form.title = "";
                 this.Form.image = "";
                 this.Form.desc = "";
@@ -262,6 +277,7 @@ export default {
               this.Form.id = "";
               this.previewImage = "";
               this.is_edit = false;
+              this.spinner = false;
             },
             // Default End
             resultListLimited(d){
